fix: refactor and update header components
