+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import '../assets/styles/globals.css';
 import { Button, ButtonScrollToUp } from '@/components';
-import { Footer, Header, HeaderFixed } from '@/modules';
+import { Footer, Header } from '@/modules';
 
 export const metadata: Metadata = {
   title: 'Студія PlanB',
@@ -17,7 +17,6 @@ export default function RootLayout({
     <html lang="ua">
       <body>
         <Header />
-        {/* <HeaderFixed /> */}
         <main>
           {children}
           <ButtonScrollToUp />
