+++ src/components/SocialMedia/socials.json
@@ -1,7 +1,7 @@
 [
   {
     "name": "instagram",
-    "href": "#"
+    "href": "https://www.instagram.com/plan_b_kp/"
   },
   {
     "name": "tic_tok",
