+++ src/modules/Header/components/HeaderFixed.tsx
@@ -1,0 +1,24 @@
+'use client';
+import React from 'react';
+import { Nav } from '.';
+import { Logo } from '@/components';
+import { Promotion, SocialMedia } from '@/components/SocialMedia/index';
+import useVisibility from '@/hooks/useVisibility';
+
+export const HeaderFixed = () => {
+  const isVisible = useVisibility(50);
+
+  return (
+    <header
+      className={`flex justify-between items-center sticky top-0 left-0 px-16 py-4 bg-stone-50 bg-header text-stone-900  z-50 border border-solid border-stone-200
+      ${isVisible ? 'visible' : 'hidden'}`}
+    >
+      <Logo className="w-[64px] h-[54px] md:w-[126px] md:h-[100px]" />
+      <Nav isMobile={false} />
+      <div className="flex flex-col gap-3 max-w-[168px]">
+        <Promotion />
+        <SocialMedia />
+      </div>
+    </header>
+  );
+};
