+++ src/modules/index.ts
@@ -1,5 +1,4 @@
 export { default as Header } from './Header/Header';
-export { default as HeaderFixed } from './Header/HeaderFixed';
 export { default as Hero } from './Hero/Hero';
 export { default as AboutUs } from './AboutUs/AboutUs';
 export { default as Services } from './Services/Services';
